+++ lib/widgets/get_images.dart
@@ -8,7 +8,7 @@ import 'loading.dart';
 
 
 class GetImage extends StatelessWidget {
-  final String imagePath;
+  final String? imagePath;
   final double width;
   final double height;
   final BoxFit fit;
@@ -17,31 +17,30 @@ class GetImage extends StatelessWidget {
   final Color? imageColor, loadingColor;
   final BorderRadius? borderRadius;
   final GestureTapCallback? onTap;
-  const GetImage({Key? key, this.imagePath = '', this.width = Siz.profileImageSize, this.height = Siz.profileImageSize, this.fit = BoxFit.cover, this.radius, this.isAssets = false, this.imageColor, this.borderRadius, this.loadingColor, this.onTap}) : super(key: key);
+  const GetImage({Key? key, this.imagePath, this.width = Siz.profileImageSize, this.height = Siz.profileImageSize, this.fit = BoxFit.cover, this.radius, this.isAssets = false, this.imageColor, this.borderRadius, this.loadingColor, this.onTap}) : super(key: key);
   static String defaultImage = 'assets/default.png';
   static bool defImageIsAsset = true;
   @override
   Widget build(BuildContext context) {
-    String imageUrl = imagePath.isEmpty?defaultImage:imagePath;
     bool isAsset = isAssets;
-    if(imagePath.isEmpty){
-      isAsset = defImageIsAsset;
+    if(imagePath == null){
+      isAsset = true;
     }
     return GestureDetector(
-      onTap: onTap??(){
-        pSetRout(page: ()=> FullPhotoView(images: [imageUrl], isAsset: isAsset,));
-      },
+      onTap: onTap??(Static.defaultImageClick?(){
+        pSetRout(page: ()=> FullPhotoView(images: [imagePath??defaultImage], isAsset: isAsset,));
+      }:null),
       child: ClipRRect(
         borderRadius: borderRadius??BorderRadius.circular(radius??Siz.defaultRadius),
         child: isAsset
             ?
         SizedBox(
           height: height, width: width,
-          child: Image.asset(imageUrl, color: imageColor, width: width, height: height, fit: fit,),
+          child: Image.asset(imagePath??defaultImage, color: imageColor, width: width, height: height, fit: fit,),
         )
             :
         CachedNetworkImage(
-          imageUrl: imageUrl,
+          imageUrl: imagePath??defaultImage,
           imageBuilder: (context, imageProvider) => Container(
             height: height,
             width: width,
@@ -62,7 +61,7 @@ class GetImage extends StatelessWidget {
                 decoration: pBoxDecoration(
                   borderRadius: borderRadius??BorderRadius.circular(radius??Siz.defaultRadius),
                 ),
-                child: const Icon(Icons.error),
+                child: Image.asset(defaultImage),
               ),
         ),
       ),