+++ lib/widgets/txt.dart
@@ -3,18 +3,20 @@ import 'package:my_widgets/my_widgets.dart';
 
 import '../utils/utils.dart';
 class Txt extends StatelessWidget {
-  final String text;
+  final String? text;
   final String?  fontFamily;
   final FontWeight? fontWeight;
   final bool hasBold, hasItalic, hasUnderLine, hasLineThrough, checkOverFlow, removeHTML;
-  final Color textColor;
+  final Color? textColor;
   final TextStyle? textStyle;
-  final double? fontSize;
+  final double? fontSize, lineHeight;
   final TextOverflow? overflow;
   final int? maxLine;
   final TextAlign textAlign;
   final TextDecoration? textDecoration;
   final Function()? onTap;
+
+
 
   const Txt(
       this.text,
@@ -22,7 +24,7 @@ class Txt extends StatelessWidget {
         this.hasItalic = false,
         this.hasUnderLine = false,
         this.hasLineThrough = false,
-        this.textColor = Colors.black,
+        this.textColor,
         this.textStyle,
         this.fontSize,
         this.fontWeight,
@@ -34,6 +36,7 @@ class Txt extends StatelessWidget {
         this.textDecoration,
         this.removeHTML = false,
         this.onTap,
+        this.lineHeight,
       Key? key,}
       ) : super(key: key);
 
@@ -41,15 +44,16 @@ class Txt extends StatelessWidget {
   Widget build(BuildContext context) {
     return GestureDetector(
       onTap: onTap,
-      child: Text(removeHTML?pRemoveHtmlIfNeeded(text):text,
-        style: textStyle??TextStyle(
+      child: Text(removeHTML?pRemoveHtmlIfNeeded(text??''):text??'',
+        style: textStyle??Style.textStyle??TextStyle(
           fontWeight: hasBold?fontWeight??Static.fontWeight??FontWeight.bold:FontWeight.normal,
           fontStyle: hasItalic?FontStyle.italic:FontStyle.normal,
           decoration: textDecoration??(hasUnderLine?TextDecoration.underline: hasLineThrough?TextDecoration.lineThrough:TextDecoration.none),
-          color: textColor,
+          color: textColor??Clr.colorTxt,
           fontSize: fontSize??Static.defaultFontSize,
           overflow: overflow??(checkOverFlow?TextOverflow.ellipsis:TextOverflow.visible),
           fontFamily: fontFamily,
+          height: lineHeight,
         ),
         maxLines: maxLine,
         textAlign: textAlign,